finished solicitatie.py
this has to be one of the most ambitious things i have created in my programming career up until now. I hope to never have to touch this code ever again. But considering the evil mastermind that is Rudi he might just make me go back to this and add some kind of functionality to it.